+++ src/components/Details.tsx
@@ -1,6 +1,6 @@
-import React, { ReactNode, useEffect, useState } from 'react'
 import { Transition } from '@headlessui/react'
 import clsx from 'clsx'
+import React, { ReactNode, useEffect, useState } from 'react'
 
 interface DetailsProps {
   children: ReactNode
@@ -22,14 +22,30 @@ const Details = ({
   return (
     <>
       <button
-        className={clsx('flex items-center justify-between w-full border-2 border-dashed border-transparent focus:outline-none -ml-0.5 -top-0.5 focus-visible:border-green-500', disabled && 'opacity-25 cursor-not-allowed')}
+        className={clsx(
+          'flex items-center justify-between w-full border-2 border-transparent focus:outline-none -ml-0.5 -top-0.5 focus-visible:border-green-500',
+          disabled && 'opacity-25 cursor-not-allowed'
+        )}
         disabled={disabled}
         onClick={() => setIsOpen(!isOpen)}
       >
         <div>{summary}</div>
 
-        <svg viewBox="0 0 24 24" fill="none" stroke="currentColor" className={clsx('w-6 h-6 text-gray-500 transform transition-transform ease-in-out duration-250', isOpen && 'rotate-180')}>
-          <path strokeLinecap="round" strokeLinejoin="round" strokeWidth="2" d="M19 9l-7 7-7-7" />
+        <svg
+          viewBox="0 0 24 24"
+          fill="none"
+          stroke="currentColor"
+          className={clsx(
+            'w-6 h-6 text-gray-500 transform transition-transform ease-in-out duration-250',
+            isOpen && 'rotate-180'
+          )}
+        >
+          <path
+            strokeLinecap="round"
+            strokeLinejoin="round"
+            strokeWidth="2"
+            d="M19 9l-7 7-7-7"
+          />
         </svg>
       </button>