+++ src/components/NotSupported.tsx
@@ -4,9 +4,9 @@ import EdgeIcon from './icons/EdgeIcon'
 
 const NotSupported = () => {
   return (
-    <div className="p-6 mt-6 text-gray-200 border-2 border-gray-700 border-dashed rounded-xl">
+    <div className="p-6 mt-6 text-gray-400 bg-gray-800 border border-gray-700 rounded-xl">
       <h2 className="flex items-center space-x-4 text-xl font-bold">
-        <div className="flex items-center justify-center flex-shrink-0 w-10 h-10 rounded-full bg-gray-1000">
+        <div className="flex items-center justify-center flex-shrink-0 w-10 h-10 bg-gray-700 rounded-full">
           <svg
             className="w-6 h-6 text-gray-400"
             fill="none"
@@ -38,13 +38,13 @@ const NotSupported = () => {
             href="//www.google.com/chrome/"
             target="_blank"
             rel="noopener noreferrer"
-            className="flex items-center justify-between px-5 py-4 space-x-3 text-sm font-semibold transition-colors ease-in-out rounded-full sm:text-lg hover:bg-gray-800 focus:bg-gray-800 focus:outline-none duration-250"
+            className="flex items-center justify-between px-5 py-4 space-x-3 text-sm font-semibold transition-colors ease-in-out rounded-full group sm:text-lg hover:bg-gray-700 focus:bg-gray-700 focus:outline-none duration-250"
           >
             <div className="flex items-center space-x-3">
               <ChromeIcon className="flex-shrink-0 opacity-75 w-7 h-7 filter grayscale" />
 
               <span>Google Chrome</span>
-              <span className="inline-block px-2 py-1 text-xs font-normal bg-gray-700 rounded-full top-px">
+              <span className="inline-block px-2 py-1 text-xs font-normal transition-colors ease-in-out bg-gray-700 rounded-full group-hover:bg-gray-600 duration-250 group-focus:bg-gray-600 top-px">
                 v86+
               </span>
             </div>
@@ -69,13 +69,13 @@ const NotSupported = () => {
             href="//www.microsoft.com/edge"
             target="_blank"
             rel="noopener noreferrer"
-            className="flex items-center justify-between px-5 py-4 space-x-3 text-sm font-semibold transition-colors ease-in-out rounded-full sm:text-lg hover:bg-gray-800 focus:bg-gray-800 focus:outline-none duration-250"
+            className="flex items-center justify-between px-5 py-4 space-x-3 text-sm font-semibold transition-colors ease-in-out rounded-full group sm:text-lg hover:bg-gray-700 focus:bg-gray-700 focus:outline-none duration-250"
           >
             <div className="flex items-center space-x-3">
               <EdgeIcon className="flex-shrink-0 opacity-75 w-7 h-7 filter grayscale" />
 
               <span>Microsoft Edge</span>
-              <span className="inline-block px-2 py-1 text-xs font-normal bg-gray-700 rounded-full top-px">
+              <span className="inline-block px-2 py-1 text-xs font-normal transition-colors ease-in-out bg-gray-700 rounded-full group-hover:bg-gray-600 duration-250 group-focus:bg-gray-600 top-px">
                 v86+
               </span>
             </div>